Migrated resource to data source
