+++ tls/provider_test.go
@@ -91,5 +91,4 @@ gQDcpC/LqiMUhSvluRBitI5ZLUn7V7jT1C3AuGW1LrTlb/q8nnhu2Gi8HY0csHvH
 4ezq8b95FJv+e/oo9ztExZ2trZ0HJ4wqTIlyN9W1DZvPQTC7gtyEFY1XvCScNoPb
 UI+cIbsQJcr9i0SADwhqs7XYIcEz9jDqrdezenaXdGtXYg==
 -----END CERTIFICATE-----
-
 `
